+++ app/franchises/[slug]/page.tsx
