+++ app/franchises/page.tsx
