+++ components/player/PlayerInfo.tsx
@@ -91,7 +91,7 @@ export default async function PlayerInfo({ playerInfo }: { playerInfo }) {
               <h2 className="text-vdcWhite text-sm">
                 {isPlayerSigned ? (
                   <Link
-                    href={`/about/franchises/${playerTeam.Franchise.slug}?team=${playerTeam.tier}`}
+                    href={`/franchises/${playerTeam.Franchise.slug}?team=${playerTeam.tier}`}
                     className="hover:text-vdcRed hover:underline"
                   >
                     {playerTeam.Franchise.slug} | {playerTeam.name}
