+++ lib/queries/franchises/franchises.ts
